Removed contours functions from geometric to new module inside fmm
